+++ webapp/src/routes/_authenticated/studies/$studyId/explore/outputs/$outputId/-components/OutputMatrixViewer/Header/DownloadVariableViewButton.tsx
