+++ webapp/src/routes/_authenticated/studies/$studyId/explore/outputs/$outputId/-components/OutputMatrixViewer/Header/ResultFilters/index.tsx
@@ -38,13 +38,7 @@ function ResultFilters() {
     setDataType,
     frequency,
     setFrequency,
-    variable,
-    setVariable,
-    clusterId,
-    setClusterId,
   } = useOutputFilters();
-
-  console.log({ year, variable, clusterId });
 
   const isYearByYearMode = monteCarloMode === "mc-ind";
