+++ webapp/src/routes/_authenticated/studies/$studyId/explore/outputs/$outputId/-utils.ts
@@ -13,69 +13,21 @@
  */
 
 import type { ListViewItem } from "@/components/page/list/ListView";
-import type { Output } from "@/services/api/studies/outputs/types";
-import type {
-  AreaVariablesDTO,
-  RenewableClusterVariablesDTO,
-  ShortTermStorageVariablesDTO,
-  ThermalClusterVariablesDTO,
-  VariablesListDTO,
-  VariableViewParams,
-} from "@/services/api/studies/outputs/variableViews/types";
 import type { AreaWithId, District, LinkElement } from "@/types/types";
+
+////////////////////////////////////////////////////////////////
+// Types
+////////////////////////////////////////////////////////////////
 
 export type ListType = "areas" | "links" | "synthesis";
 export type DataType = "values" | "details" | "details-res" | "id" | "details-STstorage";
 export type Frequency = "hourly" | "daily" | "weekly" | "monthly" | "annual";
 export type MonteCarloMode = "mc-ind" | "mc-all" | "variable-per-variable";
-
 export type Item = AreaWithId | District | LinkElement;
 
-export const MAX_YEAR = 99999;
-
-export function isArea(item: Item): item is AreaWithId {
-  return "links" in item;
-}
-
-export function isDistrict(item: Item): item is District {
-  return "addAreas" in item;
-}
-
-export function isAreaOrDistrict(item: Item): item is AreaWithId | District {
-  return isArea(item) || isDistrict(item);
-}
-
-export function isLink(item: Item): item is LinkElement {
-  return "area1" in item;
-}
-
-interface CreateOutputDataPathParams {
-  output: Output;
-  item: Item;
-  dataType: DataType;
-  frequency: Frequency;
-  year?: number;
-}
-
-export function createOutputDataPath({
-  output,
-  item,
-  dataType,
-  frequency,
-  year,
-}: CreateOutputDataPathParams): string {
-  const { id, mode = "economy" } = output;
-  const isYearPeriod = year && year > 0;
-  const periodFolder = isYearPeriod
-    ? `mc-ind/${Math.min(year, output.nbYears || MAX_YEAR)
-        .toString()
-        .padStart(5, "0")}`
-    : "mc-all";
-  const itemType = isLink(item) ? "links" : "areas";
-  const itemFolder = isLink(item) ? `${item.area1}/${item.area2}` : item.id;
-
-  return `output/${id}/${mode.toLowerCase()}/${periodFolder}/${itemType}/${itemFolder}/${dataType}-${frequency}`;
-}
+////////////////////////////////////////////////////////////////
+// Constants
+////////////////////////////////////////////////////////////////
 
 export const SYNTHESIS_ITEMS = [
   {
@@ -96,339 +48,22 @@ export const SYNTHESIS_ITEMS = [
   },
 ] as const satisfies ListViewItem[];
 
-export function matchesSearchTerm(text: string, searchTerm: string): boolean {
-  const searchTerms = searchTerm.split("|").map((term) => term.trim().toLowerCase());
-  return searchTerms.some((term) => text.toLowerCase().includes(term));
-}
-
 ////////////////////////////////////////////////////////////////
-// Variable per variable utils
+// Functions
 ////////////////////////////////////////////////////////////////
 
-/**
- * Builds parameters for variable view API requests based on item type and data type
- *
- * @param dataType - The type of data being requested (values, details, details-res, details-STstorage, or id)
- * @param selectedClusterId - The ID of the selected cluster (for cluster-specific data types)
- * @param selectedItemId - The ID of the selected item (area name or link ID)
- * @param selectedItem - The full item object (Area or LinkElement)
- * @param selectedVariable - The name of the variable to retrieve
- * @param frequency - The time frequency for the data (hourly, daily, weekly, monthly, annual)
- * @returns VariableViewParams object configured for the appropriate endpoint
- *
- * @example
- * // For area values
- * buildVariableViewParams("areas", "values", "", "area1", areaObj, "LOAD", "hourly")
- * // Returns: { type: "area", variableName: "LOAD", frequency: "hourly", areaId: "area1" }
- *
- * // For thermal cluster details
- * buildVariableViewParams("areas", "details", "cluster1", "area1", areaObj, "production", "daily")
- * // Returns: { type: "thermal", variableName: "production", frequency: "daily", areaId: "area1", clusterId: "cluster1" }
- *
- * // For links
- * buildVariableViewParams("links", "values", "", "link1", linkObj, "FLOW LIN.", "monthly")
- * // Returns: { type: "link", variableName: "FLOW LIN.", frequency: "monthly", areaFromId: "area1", areaToId: "area2" }
- */
-export function buildVariableViewParams(
-  dataType: string,
-  selectedClusterId: string,
-  selectedItem: Item,
-  selectedVariable: string,
-  frequency: Frequency,
-): VariableViewParams {
-  if (isAreaOrDistrict(selectedItem)) {
-    // Cluster/storage params
-    if (dataType === "details" && selectedClusterId) {
-      return {
-        type: "thermal",
-        variableName: selectedVariable,
-        frequency: frequency,
-        areaId: selectedItem.id,
-        clusterId: selectedClusterId,
-      };
-    }
-    if (dataType === "details-res" && selectedClusterId) {
-      return {
-        type: "renewable",
-        variableName: selectedVariable,
-        frequency: frequency,
-        areaId: selectedItem.id,
-        clusterId: selectedClusterId,
-      };
-    }
-    if (dataType === "details-STstorage" && selectedClusterId) {
-      return {
-        type: "st_storage",
-        variableName: selectedVariable,
-        frequency: frequency,
-        areaId: selectedItem.id,
-        clusterId: selectedClusterId,
-      };
-    }
-
-    // Area params
-    return {
-      type: "area",
-      variableName: selectedVariable,
-      frequency: frequency,
-      areaId: selectedItem.id,
-    };
-  }
-
-  // Link params
-  return {
-    type: "link",
-    variableName: selectedVariable,
-    frequency: frequency,
-    areaFromId: (selectedItem as LinkElement).area1,
-    areaToId: (selectedItem as LinkElement).area2,
-  };
+export function isArea(item: Item): item is AreaWithId {
+  return "links" in item;
 }
 
-/**
- * Retrieves the first variable for a selected item (area or link)
- *
- * @param variablesMetadata - The metadata containing all variables information, or null if not loaded
- * @param selectedItem - The selected item
- * @returns The first variable name for the item, or empty string if not found or no variables exist
- *
- * @example
- * // For an area
- * getFirstVariableForItem(metadata, "areas", "area1") // Returns first variable of area1
- *
- * // For a link
- * getFirstVariableForItem(metadata, "links", "area1%area2") // Returns first variable of the link
- */
-export function getFirstVariableForItem(
-  variablesMetadata: VariablesListDTO | null,
-  selectedItem: Item,
-): string {
-  if (!variablesMetadata) {
-    return "";
-  }
-
-  const data = variablesMetadata.mcInd;
-
-  if (isAreaOrDistrict(selectedItem)) {
-    const area = data.areas.find((area) => area.name === selectedItem.id);
-    return area?.variables[0] || "";
-  }
-
-  if (isLink(selectedItem)) {
-    const link = data.links.find((link) =>
-      isLinkMatch(link.area1Name, link.area2Name, selectedItem.id),
-    );
-    return link?.variables[0] || "";
-  }
-
-  return "";
+export function isDistrict(item: Item): item is District {
+  return "addAreas" in item;
 }
 
-/**
- * Extracts variables from an area based on the data type
- *
- * @param area - The area containing variables and clusters
- * @param dataType - The type of data to extract (values, details, details-res, details-STstorage)
- * @returns Array of variable names
- */
-export function getAreaVariables(area: AreaVariablesDTO, dataType: DataType): string[] {
-  switch (dataType) {
-    case "values":
-      return area.variables;
-
-    case "details":
-      return (
-        area.thermalClusters?.flatMap((cluster: ThermalClusterVariablesDTO) => cluster.variables) ||
-        []
-      );
-
-    case "details-res":
-      return (
-        area.renewableClusters?.flatMap(
-          (cluster: RenewableClusterVariablesDTO) => cluster.variables,
-        ) || []
-      );
-
-    case "details-STstorage":
-      return (
-        area.shortTermStorages?.flatMap(
-          (storage: ShortTermStorageVariablesDTO) => storage.variables,
-        ) || []
-      );
-
-    default:
-      return [];
-  }
+export function isAreaOrDistrict(item: Item): item is AreaWithId | District {
+  return isArea(item) || isDistrict(item);
 }
 
-/**
- * Checks if a link matches the selected ID (bidirectional match)
- *
- * @param area1 - First area name
- * @param area2 - Second area name
- * @param selectedId - The selected link ID to match against
- * @returns True if the link matches in either direction
- */
-function isLinkMatch(area1: string, area2: string, selectedId: string): boolean {
-  const linkId1 = `${area1}%${area2}`;
-  const linkId2 = `${area2}%${area1}`;
-  return linkId1 === selectedId || linkId2 === selectedId;
-}
-
-/**
- * Extracts variables based on item type (areas or links)
- *
- * @param variablesMetadata - The metadata containing all variables information
- * @param selectedItem - The selected item
- * @param dataType - The type of data to extract
- * @returns Array of variable names
- */
-export function getVariables(
-  variablesMetadata: VariablesListDTO,
-  selectedItem: Item,
-  dataType: DataType,
-): string[] {
-  const data = variablesMetadata.mcInd;
-
-  if (isAreaOrDistrict(selectedItem)) {
-    const area = data.areas.find((a) => a.name === selectedItem.id);
-    return area ? getAreaVariables(area, dataType) : [];
-  }
-
-  if (isLink(selectedItem)) {
-    const link = data.links.find((link) =>
-      isLinkMatch(link.area1Name, link.area2Name, selectedItem.id),
-    );
-    return link?.variables || [];
-  }
-
-  return [];
-}
-
-////////////////////////////////////////////////////////////////
-// Cluster utils
-////////////////////////////////////////////////////////////////
-
-export interface ClusterOption {
-  name: string;
-  variables: string[];
-}
-
-/**
- * Extracts thermal cluster list for a given area
- *
- * @param variablesMetadata - The metadata containing all variables information
- * @param areaId - The ID of the area
- * @returns Array of thermal clusters with their names and variables
- */
-export function getThermalClusters(
-  variablesMetadata: VariablesListDTO | null,
-  areaId: string,
-): ClusterOption[] {
-  if (!variablesMetadata || !areaId) {
-    return [];
-  }
-
-  const area = variablesMetadata.mcInd.areas.find((a) => a.name === areaId);
-  return area?.thermalClusters || [];
-}
-
-/**
- * Extracts renewable cluster list for a given area
- *
- * @param variablesMetadata - The metadata containing all variables information
- * @param areaId - The ID of the area
- * @returns Array of renewable clusters with their names and variables
- */
-export function getRenewableClusters(
-  variablesMetadata: VariablesListDTO | null,
-  areaId: string,
-): ClusterOption[] {
-  if (!variablesMetadata || !areaId) {
-    return [];
-  }
-
-  const area = variablesMetadata.mcInd.areas.find((a) => a.name === areaId);
-  return area?.renewableClusters || [];
-}
-
-/**
- * Extracts short-term storage list for a given area
- *
- * @param variablesMetadata - The metadata containing all variables information
- * @param areaId - The ID of the area
- * @returns Array of short-term storages with their names and variables
- */
-export function getShortTermStorages(
-  variablesMetadata: VariablesListDTO | null,
-  areaId: string,
-): ClusterOption[] {
-  if (!variablesMetadata || !areaId) {
-    return [];
-  }
-
-  const area = variablesMetadata.mcInd.areas.find((a) => a.name === areaId);
-  return area?.shortTermStorages || [];
-}
-
-/**
- * Gets cluster options based on the data type
- *
- * @param variablesMetadata - The metadata containing all variables information
- * @param areaId - The ID of the area
- * @param dataType - The type of data (details, details-res, details-STstorage)
- * @returns Array of cluster options based on data type
- */
-export function getClusters(
-  variablesMetadata: VariablesListDTO | null,
-  areaId: string,
-  dataType: DataType,
-): ClusterOption[] {
-  switch (dataType) {
-    case "details":
-      return getThermalClusters(variablesMetadata, areaId);
-    case "details-res":
-      return getRenewableClusters(variablesMetadata, areaId);
-    case "details-STstorage":
-      return getShortTermStorages(variablesMetadata, areaId);
-    default:
-      return [];
-  }
-}
-
-/**
- * Gets the first cluster ID for a given area and data type
- *
- * @param variablesMetadata - The metadata containing all variables information
- * @param areaId - The ID of the area
- * @param dataType - The type of data
- * @returns The first cluster ID, or empty string if not found
- */
-export function getFirstClusterId(
-  variablesMetadata: VariablesListDTO | null,
-  areaId: string,
-  dataType: DataType,
-): string {
-  const clusters = getClusters(variablesMetadata, areaId, dataType);
-  return clusters[0]?.name || "";
-}
-
-/**
- * Gets variables for a specific cluster
- *
- * @param variablesMetadata - The metadata containing all variables information
- * @param areaId - The ID of the area
- * @param dataType - The type of data
- * @param clusterId - The ID of the cluster
- * @returns Array of variable names for the cluster
- */
-export function getClusterVariables(
-  variablesMetadata: VariablesListDTO | null,
-  areaId: string,
-  dataType: DataType,
-  clusterId: string,
-): string[] {
-  const clusters = getClusters(variablesMetadata, areaId, dataType);
-  const cluster = clusters.find((c) => c.name === clusterId);
-  return cluster?.variables || [];
+export function isLink(item: Item): item is LinkElement {
+  return "area1" in item;
 }
