+++ webapp/src/services/api/studies/outputs/variableViews/index.ts
@@ -34,47 +34,7 @@ export async function getVariablesList({ studyId, outputId }: GetVariablesListPa
   const { data } = await client.get<VariablesListDTO>(
     `/v1/studies/${studyId}/output/${outputId}/variables-list`,
   );
-  // return data;
-
-  return {
-    mcInd: {
-      areas: [
-        {
-          name: "areain-1",
-          variables: ["A. DEMAND", "A. GEN", "Z. LOAD"],
-          thermalClusters: [
-            {
-              name: "cluster in2",
-              variables: ["TC1. GEN", "TC1. COST"],
-            },
-          ],
-          renewableClusters: [
-            {
-              name: "ren1",
-              variables: ["RC1. GEN", "RC1. COST"],
-            },
-          ],
-          shortTermStorages: [
-            {
-              name: "stor1",
-              variables: ["STS1. GEN", "STS1. COST"],
-            },
-          ],
-        },
-      ],
-      links: [
-        {
-          area1Name: "areain-1",
-          area2Name: "areain-2",
-          variables: ["L. FLOW", "L. LOSS"],
-        },
-      ],
-    },
-    mcAll: {
-      areas: [],
-      links: [],
-    },
-  } satisfies VariablesListDTO;
+  return data;
 }
 
 ////////////////////////////////////////////////////////////////
